requirements.txt
fastapi==0.115.14
pydantic==2.11.7
uvicorn==0.35.0
python-dotenv==1.1.1
openai==1.93.0
scikit-learn==1.7.0
hdbscan==0.8.40
numpy==2.3.1
scipy==1.16.0 